requirements.txt
google-cloud-storage
requests
pillow
celery
aiohttp
scikit-image
plotly
pandas
matplotlib
kaleido